requirements.txt
Flask
Flask-SQLAlchemy
psycopg2-binary
bcrypt
pycryptodome
flask_login
pytest